# src/AdsetManagementWeb/src/app/api/models/ImageResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ImageResponse = {
    id?: number;
    fileName?: string | null;
    contentType?: string | null;
    imageUrl?: string | null;
    uploadDate?: string;
};