# src/AdsetManagementWeb/src/app/api/models/OtherOptionsRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OtherOptionsRequest = {
    arCondicionado?: boolean | null;
    alarme?: boolean | null;
    airbag?: boolean | null;
    abs?: boolean | null;
};