# src/AdsetManagementWeb/src/app/api/models/OtherOptionsResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type OtherOptionsResponse = {
    arCondicionado?: boolean | null;
    alarme?: boolean | null;
    airbag?: boolean | null;
    abs?: boolean | null;
};