# src/AdsetManagementWeb/src/app/api/models/UpdatePacotesRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type UpdatePacotesRequest = {
    pacoteICarros?: string | null;
    pacoteWebMotors?: string | null;
};